test(mon-pix): rename password related routes folder for better discoverability
